refactor(content): extract post filter/selection toolbar into its own file
content-browser.tsx is a 3000+ line God component. PostFilterBar,
PostSelectionToolbar, and their private helpers (FilterChipTrigger,
OptionCount, FilterClearButton, SelectAllControl) only read their own
props and share no closures with the rest of the file, so they move
byte-identical into post-toolbar.tsx alongside the PostSort type export
needed to type them.
